requirements.txt
fastapi==0.111.0
uvicorn==0.22.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.7
selenium==5.9.0
openai==1.33.0
python-dotenv==1.0.0